Move shared data impls to new private module
